feat(workflow): refactor temporal client usage and introduce workflow runtime management
