+++ backend/internal/workflow/runtime/runtime.go
@@ -1,4 +1,4 @@
-package bootstrap
+package runtime
 
 import (
 	"context"
